recent changes to the css and naming , also added images
