+++ src/App.css
@@ -1,4 +1,4 @@
-.header__wrapper {
+/*.header__wrapper {
       margin: auto;
 }
 
@@ -86,7 +86,7 @@
 .image__box {
   width: 25vw;
  min-height: 65vh;
-  /* padding: 3rem; */
+
   margin-top: 3rem;
   margin-inline: auto;
   background-color: rgb(0 0 0 / 25%);
@@ -94,11 +94,11 @@
   position: absolute;
   top: 3rem;
   right: 16rem;
-}
+}*/
 
 
 /*----Main Section--------*/
-.main__section {
+/* .main__section {
   margin-top: 4.5rem;
   background-color: #e9ecef;
   padding: 2.2rem 2rem;
@@ -177,7 +177,7 @@
   position: absolute;
   top: 820px;
   right: 23%;
-}
+} */
 
 
 /*------------Section Design------*/
