+++ src/components/talentheader/header.css
@@ -1,0 +1,178 @@
+.header__wrapper {
+    margin: auto;
+}
+
+.header-section {
+   display: flex;
+   justify-content: space-around;
+   align-items: center;
+   margin: auto;
+}
+
+.header-text {
+ padding: 3rem 0;
+ margin-left: 1rem;
+ margin-bottom: 3rem;
+}
+
+.header-bold {
+text-transform: uppercase;
+font-size: 2rem;
+font-weight: 700;
+}
+
+.header-soft {
+   font-size: .85rem;
+}
+
+.btn1 {
+    background-color: #D00000;
+    border: none;
+    border-radius: 5px;
+    padding: .55rem 1rem;     
+    color: #fff;
+    font-size: .85rem;
+    font-weight: 600;
+}
+
+.header__number {
+   display: flex;
+   justify-content: space-around;
+}
+
+.number {
+   font-size: 4rem;
+   font-weight: 700;
+}
+
+.number-text {
+ font-weight: 700;
+ font-size: .85rem;
+}
+
+.plus {
+   color: #D00000;
+   font-size: 3.5rem;
+   font-weight: 700;
+}
+
+
+.right--top--ring {
+   width: 300px;
+   height: 300px;
+   border-radius: 240px;
+   border: 40px solid #eee;
+   padding: 1rem;
+   background-color: transparent;
+   position: relative;
+   left: 26rem;
+   top: -5rem;
+   z-index: -1;
+}
+
+.left--bottom--ring {
+   width: 300px;
+   height: 300px;
+   border-radius: 240px;
+   border: 40px solid #eee;
+   padding: 1rem;
+   background-color: transparent;
+   position: relative;
+   top: 10rem;
+   z-index: -1;
+}
+
+
+.image__box {
+width: 22vw;
+margin-top: 0rem;
+margin-inline: auto;
+background-color: rgb(0 0 0 / 25%);
+border-radius: 15px;
+position: absolute;
+top: 3rem;
+right: 20rem;
+}
+
+/*--------Talent Main Layout---------*/
+.main__section {
+   margin-top: 4.5rem;
+   background-color: #e9ecef;
+   padding: 2.2rem 2rem;
+ }
+ 
+ 
+ .main__flex {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+ }
+ 
+ .section-title {
+     text-align: center;
+     font-size: 1.5rem;
+     font-weight: 700;
+     padding: 1rem;
+ }
+ 
+ .col-one {
+   padding: 1.5rem 2rem;
+ }
+ 
+ .main__title {
+   font-size: 1.2rem;
+   font-weight: 700;
+   text-align: center;
+   margin-top: .85rem;
+ }
+ 
+ .main__p {
+    text-align: center;
+    padding: 0 3rem;
+ }
+ 
+ .circle {
+     width: 60px;
+     height: 60px;
+     border-radius: 150px;
+     padding: 1rem;
+     background-color: #dee2e6;
+     color: rgb(0 0 0 / 25%);
+     text-align: center;
+     font-size: 1.2rem;
+     font-weight: 700;
+     margin: auto;
+ }
+ 
+ .line1 {
+      width: 300px;
+      height: 1px;
+      background-color: #D00000;
+      position: absolute;
+      bottom: 12px;
+      left: 29.3%;
+ }
+ 
+ .arrow1 {
+   color: #D00000;
+   position: absolute;
+   top: 820px;
+   left: 46%;
+ }
+ 
+ .line2 {
+   width: 300px;
+   height: 1px;
+   background-color: #D00000;
+   position: absolute;
+   bottom: 12px;
+   left: 59.6%;
+ }
+ 
+ .arrow2 {
+   color: #D00000;
+   position: absolute;
+   top: 820px;
+   right: 23%;
+ }
+
